Split date and time tests in to separated files
